+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/ui_components/pin_code_keyboard.dart
@@ -55,6 +55,7 @@ class PinCodeKeyboard extends StatelessWidget {
     this.localizedReason,
     this.addDependencies = true,
     this.onError,
+    this.autoPromptBiometric = false,
     super.key,
   });
 
@@ -105,6 +106,12 @@ class PinCodeKeyboard extends StatelessWidget {
   /// [onError] is optional function that enable error handling out of the package
   final Function(Object error, String translatedError)? onError;
 
+  /// If set to true the biometric authentication will be triggered automatically
+  /// when the widget is loaded, requires [biometricsLocalDataSource] to be provided
+  /// otherwise the biometric authentication will be disabled
+  /// Default to false
+  final bool autoPromptBiometric;
+
   static const String _enterPinWithBiometrics =
       'Enter your pin code by authenticating with biometrics';
 
@@ -137,6 +144,7 @@ class PinCodeKeyboard extends StatelessWidget {
               biometricsLocalDataSource: biometricsLocalDataSource,
               biometricsAuthDataSource: biometricsAuthDataSource,
               localizedReason: localizedReason ?? _enterPinWithBiometrics,
+              autoPromptBiometric: autoPromptBiometric,
             ).providers,
           ],
           child: child,